msm: mdss: rotator: replace work_pending with work_busy
As per the current implementation, rotator supports asynchronous commits.
This asynchronous behavior is achieved by performing rotation in a
separate workqueue and releasing client context. However if client queues
next commit before previous one is finished then it needs to holds the
client context until previous rotation is finished. Currently to check if
previous rotation is on-going or not, driver uses work_pending API.
However this API returns true only if previous rotation job has been
queued and it is not running. If the previous job is running then it will
return false which leads driver to assume that previous job is finished,
which is incorrect and leads to wait4pingpong timeout. Fix this by using
correct API, work_busy, for rotator driver's use.

Change-Id: Ic874dabf4613385a0713a406f19f9f6379b79e11
Signed-off-by: Ujwal Patel <ujwalp@codeaurora.org>